+++ config/config.go
@@ -209,10 +209,11 @@ type autoTag struct {
 
 // SMTPRelayConfigStruct struct for parsing yaml & storing variables
 type SMTPRelayConfigStruct struct {
-	Host                    string         `yaml:"host"`
-	Port                    int            `yaml:"port"`
-	STARTTLS                bool           `yaml:"starttls"`
-	AllowInsecure           bool           `yaml:"allow-insecure"`
+	Host                    string         `yaml:"host"`               // SMTP host
+	Port                    int            `yaml:"port"`               // SMTP port
+	STARTTLS                bool           `yaml:"starttls"`           // whether to use STARTTLS
+	TLS                     bool           `yaml:"tls"`                // whether to use TLS
+	AllowInsecure           bool           `yaml:"allow-insecure"`     // allow insecure authentication, ignore TLS validation
 	Auth                    string         `yaml:"auth"`               // none, plain, login, cram-md5
 	Username                string         `yaml:"username"`           // plain & cram-md5
 	Password                string         `yaml:"password"`           // plain
@@ -234,7 +235,8 @@ type SMTPForwardConfigStruct struct {
 	Host          string `yaml:"host"`           // SMTP host
 	Port          int    `yaml:"port"`           // SMTP port
 	STARTTLS      bool   `yaml:"starttls"`       // whether to use STARTTLS
-	AllowInsecure bool   `yaml:"allow-insecure"` // allow insecure authentication
+	TLS           bool   `yaml:"tls"`            // whether to use TLS
+	AllowInsecure bool   `yaml:"allow-insecure"` // allow insecure authentication, ignore TLS validation
 	Auth          string `yaml:"auth"`           // none, plain, login, cram-md5
 	Username      string `yaml:"username"`       // plain & cram-md5
 	Password      string `yaml:"password"`       // plain
@@ -319,7 +321,7 @@ func VerifyConfig() error {
 	}
 
 	if SMTPTLSCert != "" && SMTPTLSKey == "" || SMTPTLSCert == "" && SMTPTLSKey != "" {
-		return errors.New("[smtp] You must provide both an SMTP TLS certificate and a key")
+		return errors.New("[smtp] you must provide both an SMTP TLS certificate and a key")
 	}
 
 	if SMTPTLSCert != "" {
@@ -404,7 +406,7 @@ func VerifyConfig() error {
 		}
 	}
 	if POP3TLSCert != "" && POP3TLSKey == "" || POP3TLSCert == "" && POP3TLSKey != "" {
-		return errors.New("[pop3] You must provide both a POP3 TLS certificate and a key")
+		return errors.New("[pop3] you must provide both a POP3 TLS certificate and a key")
 	}
 	if POP3Listen != "" {
 		_, err := net.ResolveTCPAddr("tcp", POP3Listen)